+++ routes/routes.go
@@ -1,6 +1,7 @@
 package routes
 
 import (
+	"fmt"
 	"os"
 	"time"
 
@@ -9,14 +10,19 @@ import (
 	"gitlab.com/nunet/device-management-service/firecracker"
 	"gitlab.com/nunet/device-management-service/firecracker/telemetry"
 	"gitlab.com/nunet/device-management-service/integrations/tokenomics"
+	"gitlab.com/nunet/device-management-service/internal/tracing"
 	"gitlab.com/nunet/device-management-service/libp2p"
 	"gitlab.com/nunet/device-management-service/libp2p/machines"
 	"gitlab.com/nunet/device-management-service/onboarding"
+	"go.opentelemetry.io/contrib/instrumentation/github.com/gin-gonic/gin/otelgin"
 )
 
 func SetupRouter() *gin.Engine {
 	router := gin.Default()
 	router.Use(cors.New(getCustomCorsConfig()))
+
+	fmt.Println("-------------------------------", tracing.ServiceName)
+	router.Use(otelgin.Middleware(tracing.ServiceName))
 
 	v1 := router.Group("/api/v1")